// src/types/events.ts
/* istanbul ignore file */

// eslint-disable-next-line no-shadow
export enum InternalEventTypes {
    DISCORD_MESSAGE = 'DISCORD_MESSAGE',

    MONITOR_STATE_CHANGE = 'MONITOR_STATE_CHANGE',

    LOG_ENTRY = 'LOG_ENTRY',
    METRIC_ENTRY = 'METRIC_ENTRY',

    GAME_UPDATED = 'GAME_UPDATED',
    MOD_UPDATED = 'MOD_UPDATED',

    INTERNAL_MOD_INSTALL = 'INTERNAL_MOD_INSTALL',
    GET_INTERNAL_MODS = 'GET_INTERNAL_MODS',
}